aegis_multimodal_ai_system/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
requests
gradio
spacy
